+++ cmd/acm.go
@@ -1,4 +1,4 @@
-package command
+package cmd
 
 import (
 	"fmt"
@@ -7,12 +7,13 @@ import (
 	"github.com/urfave/cli"
 )
 
-func ListEc2(c *cli.Context) error {
+func ListCertificates(c *cli.Context) error {
 	profile := c.GlobalString("profile")
 	region := c.GlobalString("region")
-	tag := c.String("tag")
 
-	err := aws.DescribeInstances(profile, region, tag)
+	fmt.Printf("\x1b[32mAWS_PROFILE: %s , REGION: %s\x1b[0m\n", profile, region)
+
+	err := aws.ListCertificates(profile, region)
 	if err != nil {
 		return fmt.Errorf("%v", err)
 	}