+++ cmd/elbv2.go
@@ -1,4 +1,4 @@
-package command
+package cmd
 
 import (
 	"fmt"
@@ -7,11 +7,13 @@ import (
 	"github.com/urfave/cli"
 )
 
-func ListHostedZones(c *cli.Context) error {
+func ListElbv2(c *cli.Context) error {
 	profile := c.GlobalString("profile")
 	region := c.GlobalString("region")
 
-	err := aws.ListHostedZones(profile, region)
+	fmt.Printf("\x1b[32mAWS_PROFILE: %s , REGION: %s\x1b[0m\n", profile, region)
+
+	err := aws.DescribeLoadBalancersv2(profile, region)
 	if err != nil {
 		return fmt.Errorf("%v", err)
 	}