+++ cmd/route53.go
@@ -1,0 +1,22 @@
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/sfuruya0612/snatch/internal/aws"
+	"github.com/urfave/cli"
+)
+
+func ListHostedZones(c *cli.Context) error {
+	profile := c.GlobalString("profile")
+	region := c.GlobalString("region")
+
+	fmt.Printf("\x1b[32mAWS_PROFILE: %s , REGION: %s\x1b[0m\n", profile, region)
+
+	err := aws.ListHostedZones(profile, region)
+	if err != nil {
+		return fmt.Errorf("%v", err)
+	}
+
+	return nil
+}