+++ cmd/ssm.go
@@ -1,4 +1,4 @@
-package command
+package cmd
 
 import (
 	"fmt"
@@ -11,6 +11,8 @@ func StartSession(c *cli.Context) error {
 	profile := c.GlobalString("profile")
 	region := c.GlobalString("region")
 
+	fmt.Printf("\x1b[32mAWS_PROFILE: %s , REGION: %s\x1b[0m\n", profile, region)
+
 	err := aws.StartSession(profile, region)
 	if err != nil {
 		return fmt.Errorf("%v", err)
@@ -22,6 +24,8 @@ func StartSession(c *cli.Context) error {
 func SendCommand(c *cli.Context) error {
 	profile := c.GlobalString("profile")
 	region := c.GlobalString("region")
+
+	fmt.Printf("\x1b[32mAWS_PROFILE: %s , REGION: %s\x1b[0m\n", profile, region)
 
 	args := c.Args()
 	file := c.String("file")