+++ internal/aws/session.go
@@ -1,8 +1,6 @@
 package aws
 
 import (
-	"fmt"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
 )
@@ -13,8 +11,6 @@ func getSession(profile string, region string) *session.Session {
 		Profile:           profile,
 		Config:            aws.Config{Region: aws.String(region)},
 	}
-
-	fmt.Printf("\x1b[32mAWS_PROFILE: %s , REGION: %s\x1b[0m\n", profile, region)
 
 	return session.Must(session.NewSessionWithOptions(opts))
 }