+++ main.go
@@ -4,9 +4,11 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/sfuruya0612/snatch/cmd/snatch/command"
+	"github.com/sfuruya0612/snatch/cmd"
 	"github.com/urfave/cli"
 )
+
+const version = "19.10.1"
 
 var (
 	date      string
@@ -26,10 +28,16 @@ func main() {
 func New(date, hash, goversion string) *cli.App {
 	app := cli.NewApp()
 
-	app.Name = "snatch"
-	app.Usage = "This is the cli command to get and display Amazon Web Services resources."
-	app.Version = fmt.Sprintf("%s %s (Build by: %s)", date, hash, goversion)
 	app.EnableBashCompletion = true
+
+	app.Name = "snatch"
+	app.Usage = "Cli command to get and display Amazon Web Services resources."
+
+	if date != "" || hash != "" || goversion != "" {
+		app.Version = fmt.Sprintf("%s %s (Build by: %s)", date, hash, goversion)
+	} else {
+		app.Version = version
+	}
 
 	app.Flags = []cli.Flag{
 		cli.StringFlag{
@@ -49,7 +57,7 @@ func New(date, hash, goversion string) *cli.App {
 		{
 			Name:   "ec2",
 			Usage:  "Get a list of EC2 resources. (API: DescribeInstances)",
-			Action: command.ListEc2,
+			Action: cmd.ListEc2,
 			Flags: []cli.Flag{
 				cli.StringFlag{
 					Name:  "tag, t",
@@ -60,44 +68,60 @@ func New(date, hash, goversion string) *cli.App {
 		{
 			Name:   "rds",
 			Usage:  "Get a list of RDS resources. (API: DescribeDbInstances)",
-			Action: command.ListRds,
+			Action: cmd.ListRds,
 		},
 		{
 			Name:   "ec",
 			Usage:  "Get a list of ElastiCache Cluster resources. (API: DescribeCacheClusters)",
-			Action: command.ListElasticache,
+			Action: cmd.ListElasticache,
 			Subcommands: []cli.Command{
 				{
 					Name:   "rg",
 					Usage:  "Get a list of ElastiCache Node resources. (API: DescribeReplicationGroups)",
-					Action: command.ListReplicationGroups,
+					Action: cmd.ListReplicationGroups,
 				},
 			},
 		},
 		{
 			Name:   "elb",
 			Usage:  "Get a list of ELB(Classic) resources. (API: DescribeLoadBalancers)",
-			Action: command.ListElb,
+			Action: cmd.ListElb,
 		},
 		{
 			Name:   "elbv2",
 			Usage:  "Get a list of ELB(Application & Network) resources. (API: DescribeLoadBalancers)",
-			Action: command.ListElbv2,
+			Action: cmd.ListElbv2,
 		},
 		{
 			Name:   "route53",
 			Usage:  "Get a list of Rotue53 Record resources. (API: ListHostedZones and ListResourceRecordSets)",
-			Action: command.ListHostedZones,
+			Action: cmd.ListHostedZones,
+		},
+		{
+			Name:   "acm",
+			Usage:  "Get a list of ACM resources. (API: ListCertificates and DescribeCertificate)",
+			Action: cmd.ListCertificates,
+		},
+		{
+			Name:   "s3",
+			Usage:  "Get Objects in selected S3 Bucket at interactive prompt. (API: ListBuckets and ListObjects)",
+			Action: cmd.ListBuckets,
+			Flags: []cli.Flag{
+				cli.BoolFlag{
+					Name:  "l",
+					Usage: "Get Objects list.",
+				},
+			},
 		},
 		{
 			Name:   "ssm",
 			Usage:  "Start a session on your instances by launching bash or shell terminal. (API: StartSession)",
-			Action: command.StartSession,
+			Action: cmd.StartSession,
 			Subcommands: []cli.Command{
 				{
 					Name:   "run",
 					Usage:  "Runs commands on one target instance. (API: SendCommand)",
-					Action: command.SendCommand,
+					Action: cmd.SendCommand,
 					Flags: []cli.Flag{
 						cli.StringFlag{
 							Name:  "file, f",
@@ -115,6 +139,23 @@ func New(date, hash, goversion string) *cli.App {
 				},
 			},
 		},
+		{
+			Name:   "logs",
+			Usage:  "Display messages for selected log groups and streams at interactive prompt. (API: )",
+			Action: cmd.DescribeLogGroups,
+			Flags: []cli.Flag{
+				cli.BoolFlag{
+					Name:  "f",
+					Usage: "Like `tail -f`.",
+				},
+			},
+		},
+		{
+			Name:   "cfn",
+			Usage:  "Display a list of stacks. (API: )",
+			Action: cmd.DescribeStacks,
+		},
 	}
+
 	return app
 }