+++ test/run_cmd.sh
@@ -1,0 +1,16 @@
+#!/bin/sh
+
+# RunCommand Test scripts
+# e.g. snatch ssm run -f test/run_cmd.sh -t Role:test
+
+# Uptime
+uptime
+
+# nginx install
+sudo apt install -y nginx > /dev/null
+
+# nginx status
+sudo systemctl status nginx
+
+# Check 80 port listen
+lsof -i:80