+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## [19.11.1] - 2019-11-01
+### Added
+- Get AWS resources describe
+  - Cloudwatch Logs
+  - S3
+  - ACM
+  - Cloudformation
+  - Testing CFn template
+- travis CI
+
+### Changed
+- EC2 tag filter
+- Start-session cursor display
+
+---
+
 ## [19.10.1] - 2019-10-09
 ### Added
 - Get AWS resources describe